Move statics for `PostStateValidators` to `SharedApiTypes`
