+++ data/beaconrestapi/src/main/java/tech/pegasys/teku/beaconrestapi/handlers/v1/beacon/PostStateValidators.java
@@ -15,11 +15,11 @@ package tech.pegasys.teku.beaconrestapi.handlers.v1.beacon;
 
 import static tech.pegasys.teku.beaconrestapi.BeaconRestApiTypes.PARAMETER_STATE_ID;
 import static tech.pegasys.teku.ethereum.json.types.beacon.StateValidatorDataBuilder.STATE_VALIDATORS_RESPONSE_TYPE;
+import static tech.pegasys.teku.ethereum.json.types.beacon.StateValidatorRequestType.STATE_VALIDATOR_REQUEST_TYPE;
 import static tech.pegasys.teku.ethereum.json.types.beacon.StatusParameter.getApplicableValidatorStatuses;
 import static tech.pegasys.teku.infrastructure.http.HttpStatusCodes.SC_BAD_REQUEST;
 import static tech.pegasys.teku.infrastructure.http.HttpStatusCodes.SC_OK;
 import static tech.pegasys.teku.infrastructure.http.RestApiConstants.TAG_BEACON;
-import static tech.pegasys.teku.infrastructure.json.types.CoreTypes.STRING_TYPE;
 
 import com.fasterxml.jackson.core.JsonParseException;
 import com.fasterxml.jackson.core.JsonProcessingException;
@@ -31,9 +31,9 @@ import tech.pegasys.teku.api.ChainDataProvider;
 import tech.pegasys.teku.api.DataProvider;
 import tech.pegasys.teku.api.response.v1.beacon.ValidatorStatus;
 import tech.pegasys.teku.ethereum.json.types.beacon.StateValidatorData;
+import tech.pegasys.teku.ethereum.json.types.beacon.StateValidatorRequestType;
 import tech.pegasys.teku.ethereum.json.types.beacon.StatusParameter;
 import tech.pegasys.teku.infrastructure.async.SafeFuture;
-import tech.pegasys.teku.infrastructure.json.types.DeserializableTypeDefinition;
 import tech.pegasys.teku.infrastructure.restapi.endpoints.AsyncApiResponse;
 import tech.pegasys.teku.infrastructure.restapi.endpoints.EndpointMetadata;
 import tech.pegasys.teku.infrastructure.restapi.endpoints.RestApiEndpoint;
@@ -41,22 +41,6 @@ import tech.pegasys.teku.infrastructure.restapi.endpoints.RestApiRequest;
 import tech.pegasys.teku.spec.datastructures.metadata.ObjectAndMetaData;
 
 public class PostStateValidators extends RestApiEndpoint {
-  private static final DeserializableTypeDefinition<RequestBody> REQUEST_TYPE =
-      DeserializableTypeDefinition.object(RequestBody.class)
-          .name("PostStateValidatorsRequestBody")
-          .initializer(RequestBody::new)
-          .withOptionalField(
-              "ids",
-              DeserializableTypeDefinition.listOf(STRING_TYPE),
-              RequestBody::getMaybeIds,
-              RequestBody::setIds)
-          .withOptionalField(
-              "statuses",
-              DeserializableTypeDefinition.listOf(STRING_TYPE),
-              RequestBody::getMaybeStringStatuses,
-              RequestBody::setStatuses)
-          .build();
-
   private final ChainDataProvider chainDataProvider;
 
   public PostStateValidators(final DataProvider dataProvider) {
@@ -71,7 +55,7 @@ public class PostStateValidators extends RestApiEndpoint {
             .description(
                 "Returns filterable list of validators with their balance, status and index.")
             .pathParam(PARAMETER_STATE_ID)
-            .requestBodyType(REQUEST_TYPE)
+            .requestBodyType(STATE_VALIDATOR_REQUEST_TYPE)
             .tags(TAG_BEACON)
             .response(SC_OK, "Request successful", STATE_VALIDATORS_RESPONSE_TYPE)
             .withNotFoundResponse()
@@ -81,7 +65,7 @@ public class PostStateValidators extends RestApiEndpoint {
 
   @Override
   public void handleRequest(RestApiRequest request) throws JsonProcessingException {
-    final Optional<RequestBody> requestBody;
+    final Optional<StateValidatorRequestType> requestBody;
 
     try {
       requestBody = request.getOptionalRequestBody();
@@ -95,9 +79,10 @@ public class PostStateValidators extends RestApiEndpoint {
       return;
     }
 
-    final List<String> validators = requestBody.map(RequestBody::getIds).orElse(List.of());
+    final List<String> validators =
+        requestBody.map(StateValidatorRequestType::getIds).orElse(List.of());
     final List<StatusParameter> statusParameters =
-        requestBody.map(RequestBody::getStatuses).orElse(List.of());
+        requestBody.map(StateValidatorRequestType::getStatuses).orElse(List.of());
 
     final Set<ValidatorStatus> statusFilter = getApplicableValidatorStatuses(statusParameters);
 
@@ -111,43 +96,5 @@ public class PostStateValidators extends RestApiEndpoint {
                 maybeData
                     .map(AsyncApiResponse::respondOk)
                     .orElseGet(AsyncApiResponse::respondNotFound)));
-  }
-
-  static class RequestBody {
-    private List<String> ids = List.of();
-    private List<StatusParameter> statuses = List.of();
-
-    RequestBody() {}
-
-    public RequestBody(final List<String> ids, final List<StatusParameter> statuses) {
-      this.ids = ids;
-      this.statuses = statuses;
-    }
-
-    public List<String> getIds() {
-      return ids;
-    }
-
-    public Optional<List<String>> getMaybeIds() {
-      return ids.isEmpty() ? Optional.empty() : Optional.of(ids);
-    }
-
-    public void setIds(final Optional<List<String>> ids) {
-      ids.ifPresent(i -> this.ids = i);
-    }
-
-    public List<StatusParameter> getStatuses() {
-      return statuses;
-    }
-
-    public Optional<List<String>> getMaybeStringStatuses() {
-      return statuses.isEmpty()
-          ? Optional.empty()
-          : Optional.of(statuses.stream().map(Enum::name).toList());
-    }
-
-    public void setStatuses(final Optional<List<String>> statuses) {
-      statuses.ifPresent(s -> this.statuses = s.stream().map(StatusParameter::parse).toList());
-    }
   }
 }
